requirements.txt
# Core
Django==5.1.4
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Celery + Redis
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Excel
openpyxl==3.1.5
pandas==2.2.3

# Notifications
requests==2.32.3

# Utils
Pillow==11.0.0
django-cors-headers==4.6.0
whitenoise==6.8.2
